Use the network client without an specific Adapter
