+++ web/src/client/network/index.js
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) [2022] SUSE LLC
+ * Copyright (c) [2023] SUSE LLC
  *
  * All Rights Reserved.
  *
@@ -21,9 +21,26 @@
 
 // @ts-check
 
+import DBusClient from "../dbus";
 import { NetworkManagerAdapter } from "./network_manager";
-import { ConnectionTypes, ConnectionState } from "./model";
-import { AgamaNetworkAdapter } from "./agama_network";
+import cockpit from "../../lib/cockpit";
+import { createConnection, ConnectionTypes, ConnectionState } from "./model";
+
+const SERVICE_NAME = "org.opensuse.Agama1";
+const CONNECTIONS_IFACE = "org.opensuse.Agama1.Network.Connections";
+const CONNECTIONS_PATH = "/org/opensuse/Agama1/Network/connections";
+const CONNECTION_IFACE = "org.opensuse.Agama1.Network.Connection";
+const CONNECTIONS_NAMESPACE = "/org/opensuse/Agama1/Network/connections";
+const IP_IFACE = "org.opensuse.Agama1.Network.Connection.IP";
+const WIRELESS_IFACE = "org.opensuse.Agama1.Network.Connection.Wireless";
+
+const DeviceType = Object.freeze({
+  LOOPBACK: 0,
+  ETHERNET: 1,
+  WIRELESS: 2,
+  DUMMY: 3,
+  BOND: 4
+});
 
 /**
  * @typedef {import("./model").NetworkSettings} NetworkSettings
@@ -78,15 +95,20 @@ const NetworkEventTypes = Object.freeze({
  */
 class NetworkClient {
   /**
-   * @param {NetworkAdapter} adapter - Network adapter. By default, it is set to
-o  *   NetworkManagerAdapter.
+   * @param {string} address - D-Bus address
    */
-  constructor(adapter) {
-    this.adapter = adapter;
-    /** @type {!boolean} */
+  constructor(address) {
     this.subscribed = false;
+    this.nm = new NetworkManagerAdapter();
+    this.client = new DBusClient(SERVICE_NAME, address);
+    this.proxies = {
+      connectionsRoot: null,
+      connections: {},
+      ipConfigs: {},
+      wireless: {}
+    };
+    this.eventsHandler = null;
     this.setUpDone = false;
-    /** @type {NetworkEventFn[]} */
     this.handlers = [];
   }
 
@@ -111,7 +133,15 @@ o  *   NetworkManagerAdapter.
   async setUp() {
     if (this.setUpDone) return;
 
-    return this.adapter.setUp(e => this.handlers.forEach(f => f(e)));
+    this.proxies = {
+      connectionsRoot: await this.client.proxy(CONNECTIONS_IFACE, CONNECTIONS_PATH),
+      connections: await this.client.proxies(CONNECTION_IFACE, CONNECTIONS_NAMESPACE),
+      ipConfigs: await this.client.proxies(IP_IFACE, CONNECTIONS_NAMESPACE),
+      wireless: await this.client.proxies(WIRELESS_IFACE, CONNECTIONS_NAMESPACE)
+    };
+
+    this.setUpDone = true;
+    return this.nm.setUp(e => this.handlers.forEach(f => f(e)));
   }
 
   /**
@@ -120,7 +150,7 @@ o  *   NetworkManagerAdapter.
    * @return {ActiveConnection[]}
    */
   activeConnections() {
-    return this.adapter.activeConnections();
+    return this.nm.activeConnections();
   }
 
   /**
@@ -129,7 +159,7 @@ o  *   NetworkManagerAdapter.
    * @return {Promise<Connection[]>}
    */
   connections() {
-    return this.adapter.connections();
+    return this.nm.connections();
   }
 
   /**
@@ -138,7 +168,7 @@ o  *   NetworkManagerAdapter.
    * @return {AccessPoint[]}
    */
   accessPoints() {
-    return this.adapter.accessPoints();
+    return this.nm.accessPoints();
   }
 
   /**
@@ -147,7 +177,7 @@ o  *   NetworkManagerAdapter.
    * @param {Connection} connection - connection to be activated
    */
   async connectTo(connection) {
-    return this.adapter.connectTo(connection);
+    return this.nm.connectTo(connection);
   }
 
   /**
@@ -157,37 +187,161 @@ o  *   NetworkManagerAdapter.
    * @param {object} options - connection options
    */
   async addAndConnectTo(ssid, options) {
-    return this.adapter.addAndConnectTo(ssid, options);
+    // duplicated code (see network manager adapter)
+    const wireless = { ssid };
+    if (options.security) wireless.security = options.security;
+    if (options.password) wireless.password = options.password;
+    if (options.hidden) wireless.hidden = options.hidden;
+
+    const connection = createConnection({
+      name: ssid,
+      wireless
+    });
+
+    const added = await this.addConnection(connection);
+    return this.connectTo(added);
   }
 
   /**
    * Adds a new connection
    *
    * @param {Connection} connection - Connection to add
+   * @return {Promise<Connection>} the added connection
    */
   async addConnection(connection) {
-    return this.adapter.addConnection(connection);
+    const { name } = connection;
+    const proxy = await this.client.proxy(CONNECTIONS_IFACE, CONNECTIONS_PATH);
+    const ctype = (connection.wireless) ? DeviceType.WIRELESS : DeviceType.ETHERNET;
+    const path = await proxy.AddConnection(name, ctype);
+    await this.updateConnectionAt(path, { ...connection, id: name });
+    return this.connectionFromPath(path);
   }
 
   /**
    * Returns the connection with the given ID
    *
-   * @param {string} id - Connection ID
+   * @param {string} uuid - Connection ID
+   * @return {Promise<Connection|undefined>}
+   */
+  async getConnection(uuid) {
+    const path = await this.getConnectionPath(uuid);
+    if (path) {
+      return this.connectionFromPath(path);
+    }
+  }
+
+  /**
+   * Returns a connection from the given D-Bus path
+   *
+   * @param {string} path - Path of the D-Bus object representing the connection
    * @return {Promise<Connection>}
    */
-  async getConnection(id) {
-    return this.adapter.getConnection(id);
+  async connectionFromPath(path) {
+    const connection = await this.proxies.connections[path];
+    const ip = await this.proxies.ipConfigs[path];
+
+    const conn = {
+      id: connection.Id,
+      uuid: connection.Uuid,
+      name: connection.Interface,
+      ipv4: {
+        method: ip.Method4,
+        nameServers: ip.Nameservers,
+        addresses: ip.Addresses.map(addr => {
+          const [address, prefix] = addr.split("/");
+          return { address, prefix };
+        }),
+        gateway: ip.Gateway4
+      },
+    };
+
+    const wireless = await this.proxies.wireless[path];
+    if (wireless) {
+      conn.wireless = {
+        ssid: window.atob(wireless.SSID),
+        hidden: false, // TODO implement the 'hidden' property
+        mode: wireless.Mode,
+        security: wireless.Security // see AgamaSecurityProtocols
+      };
+    }
+
+    return createConnection(conn);
+  }
+
+  /**
+   * Sets a property for a given path
+   *
+   * @param {string} path - Object path.
+   * @param {string} iface - Interface name.
+   * @param {string} property - Property name.
+   * @param {object} value - Property value. The value should be created by
+   * using the cockpit.variant() function.
+   */
+  async setProperty(path, iface, property, value) {
+    return this.client.call(path, "org.freedesktop.DBus.Properties", "Set", [iface, property, value]);
+  }
+
+  /**
+   * Returns the D-Bus path of the connection.
+   *
+   * @param {string} uuid - Connection UUID
+   * @return {Promise<string|undefined>} - Connection D-Bus path
+   */
+  async getConnectionPath(uuid) {
+    for (const path in this.proxies.connections) {
+      const proxy = await this.proxies.connections[path];
+      if (proxy.Uuid === uuid) {
+        return path;
+      }
+    }
   }
 
   /**
    * Updates the connection
    *
-   * It uses the 'path' to match the connection in the backend.
+   * It uses the 'uuid' to match the connection in the backend.
    *
    * @param {Connection} connection - Connection to update
+   * @return {Promise<boolean>} - the promise resolves to true if the connection
+   *   was successfully updated and to false it it does not exist.
    */
   async updateConnection(connection) {
-    return this.adapter.updateConnection(connection);
+    const path = await this.getConnectionPath(connection.uuid);
+    if (path === undefined) {
+      return false;
+    }
+
+    await this.updateConnectionAt(path, connection);
+    return true;
+  }
+
+  /**
+   * Updates the connection in the given path
+   *
+   *
+   * @param {string} path - D-Bus path of the connection to update.
+   * @param {Connection} connection - Connection to update.
+   */
+  async updateConnectionAt(path, connection) {
+    const { ipv4, wireless } = connection;
+    await this.setProperty(path, IP_IFACE, "Method4", cockpit.variant("s", ipv4.method));
+    await this.setProperty(path, IP_IFACE, "Gateway4", cockpit.variant("s", ipv4.gateway));
+    const addresses = ipv4.addresses.map(a => `${a.address}/${a.prefix}`);
+    await this.setProperty(path, IP_IFACE, "Addresses", cockpit.variant("as", addresses));
+    await this.setProperty(path, IP_IFACE, "Nameservers", cockpit.variant("as", ipv4.nameServers));
+
+    if (wireless) {
+      await this.setProperty(path, WIRELESS_IFACE, "Mode", cockpit.variant("s", "infrastructure"));
+      if (wireless.password) {
+        await this.setProperty(path, WIRELESS_IFACE, "Password", cockpit.variant("s", wireless.password));
+      }
+      await this.setProperty(path, WIRELESS_IFACE, "Security", cockpit.variant("s", wireless.security));
+      const ssid = cockpit.byte_array(wireless.ssid);
+      await this.setProperty(path, WIRELESS_IFACE, "SSID", cockpit.variant("ay", ssid));
+    }
+
+    // TODO: apply the changes only in this connection
+    return this.proxies.connectionsRoot.Apply();
   }
 
   /**
@@ -198,7 +352,7 @@ o  *   NetworkManagerAdapter.
    * @param {Connection} connection - Connection to delete
    */
   async deleteConnection(connection) {
-    return this.adapter.deleteConnection(connection);
+    return this.nm.deleteConnection(connection);
   }
 
   /*
@@ -209,7 +363,7 @@ o  *   NetworkManagerAdapter.
    * @return {Promise<IPAddress[]>}
    */
   addresses() {
-    const conns = this.adapter.activeConnections();
+    const conns = this.activeConnections();
     return conns.flatMap(c => c.addresses);
   }
 
@@ -217,15 +371,13 @@ o  *   NetworkManagerAdapter.
   * Returns network general settings
   */
   settings() {
-    return this.adapter.settings();
+    return this.nm.settings();
   }
 }
 
 export {
-  AgamaNetworkAdapter,
   ConnectionState,
   ConnectionTypes,
   NetworkClient,
-  NetworkManagerAdapter,
   NetworkEventTypes
 };
